PictureSaver was added
