+++ TagsCloud.WPF/PictureSaver/IPictureSaver.cs
@@ -1,0 +1,8 @@
+﻿using System.Windows;
+
+namespace TagsCloud.WPF.PictureSaver;
+
+public interface IPictureSaver
+{
+    public void SavePicture(object sender, RoutedEventArgs e);
+}
